requirements.txt
torch>=1.10
opencv-python>=4.5
numpy>=1.20
PyYAML>=5.3
scipy>=1.6
